+++ asilib/__init__.py
@@ -1,7 +1,7 @@
 import pathlib
 import configparser
 
-__version__ = '0.21.1'
+__version__ = '0.23.0'
 
 # Load the configuration settings.
 HERE = pathlib.Path(__file__).parent.resolve()